+++ bcos-scheduler/CMakeLists.txt
@@ -2,16 +2,14 @@ cmake_minimum_required(VERSION 3.15)
 
 project(bcos-scheduler VERSION "3.0.0")
 
-aux_source_directory(bcos-scheduler SRCS)
-
-include_directories(${CMAKE_SOURCE_DIR})
-include_directories(${CMAKE_SOURCE_DIR}/bcos-framework)
+aux_source_directory(src SRCS)
 
 find_package(TBB CONFIG REQUIRED)
 find_package(Threads REQUIRED)
 
 add_library(scheduler ${SRCS})
-target_link_libraries(scheduler framework TBB::tbb)
+target_link_libraries(scheduler PUBLIC framework TBB::tbb)
+target_include_directories(scheduler PUBLIC ${CMAKE_SOURCE_DIR}/bcos-framework)
 
 if(TESTS)
     enable_testing()
