+++ bcos-scheduler/tests/CMakeLists.txt
@@ -8,7 +8,7 @@ set(TEST_BINARY_NAME bcos-dispatcher-test)
 include_directories(${CMAKE_SOURCE_DIR}/bcos-scheduler)
 
 add_executable(${TEST_BINARY_NAME} ${SOURCES})
-target_include_directories(${TEST_BINARY_NAME} PRIVATE . ${CMAKE_SOURCE_DIR})
+target_include_directories(${TEST_BINARY_NAME} PRIVATE . ${CMAKE_SOURCE_DIR}/bcos-scheduler/src)
 target_compile_options(${TEST_BINARY_NAME} PRIVATE -Wno-error -Wno-unused-variable)
 
 hunter_add_package(wedpr-crypto)
