+++ bcos-scheduler/tests/mock/MockExecutor3.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../bcos-scheduler/Common.h"
+#include "Common.h"
 #include "MockExecutor.h"
 #include <boost/lexical_cast.hpp>
 #include <tuple>
