+++ bcos-scheduler/tests/testChecksumAddress.cpp
@@ -1,18 +1,18 @@
-#include "bcos-scheduler/ExecutorManager.h"
-#include "bcos-scheduler/SchedulerImpl.h"
-#include "bcos-scheduler/ChecksumAddress.h"
-#include "interfaces/crypto/CryptoSuite.h"
-#include "interfaces/crypto/KeyPairInterface.h"
-#include "interfaces/executor/ExecutionMessage.h"
-#include "interfaces/ledger/LedgerInterface.h"
-#include "interfaces/protocol/BlockHeaderFactory.h"
-#include "interfaces/protocol/ProtocolTypeDef.h"
-#include "interfaces/protocol/Transaction.h"
-#include "interfaces/protocol/TransactionReceipt.h"
-#include "interfaces/protocol/TransactionReceiptFactory.h"
-#include "interfaces/protocol/TransactionSubmitResult.h"
-#include "interfaces/storage/StorageInterface.h"
-#include "libprotocol/TransactionSubmitResultFactoryImpl.h"
+#include "ExecutorManager.h"
+#include "SchedulerImpl.h"
+#include "ChecksumAddress.h"
+#include "bcos-framework/interfaces/crypto/CryptoSuite.h"
+#include "bcos-framework/interfaces/crypto/KeyPairInterface.h"
+#include "bcos-framework/interfaces/executor/ExecutionMessage.h"
+#include "bcos-framework/interfaces/ledger/LedgerInterface.h"
+#include "bcos-framework/interfaces/protocol/BlockHeaderFactory.h"
+#include "bcos-framework/interfaces/protocol/ProtocolTypeDef.h"
+#include "bcos-framework/interfaces/protocol/Transaction.h"
+#include "bcos-framework/interfaces/protocol/TransactionReceipt.h"
+#include "bcos-framework/interfaces/protocol/TransactionReceiptFactory.h"
+#include "bcos-framework/interfaces/protocol/TransactionSubmitResult.h"
+#include "bcos-framework/interfaces/storage/StorageInterface.h"
+#include "bcos-framework/libprotocol/TransactionSubmitResultFactoryImpl.h"
 #include "mock/MockExecutor.h"
 #include "mock/MockExecutor3.h"
 #include "mock/MockExecutorForCall.h"
