+++ bcos-scheduler/tests/testCommitBlock.cpp
@@ -1,12 +1,12 @@
-#include "bcos-scheduler/BlockExecutive.h"
-#include "bcos-scheduler/ExecutorManager.h"
-#include "bcos-scheduler/SchedulerImpl.h"
-#include "interfaces/crypto/CryptoSuite.h"
-#include "interfaces/executor/ExecutionMessage.h"
-#include "interfaces/ledger/LedgerInterface.h"
-#include "interfaces/protocol/BlockHeaderFactory.h"
-#include "interfaces/protocol/TransactionReceiptFactory.h"
-#include "interfaces/storage/StorageInterface.h"
+#include "BlockExecutive.h"
+#include "ExecutorManager.h"
+#include "SchedulerImpl.h"
+#include "bcos-framework/interfaces/crypto/CryptoSuite.h"
+#include "bcos-framework/interfaces/executor/ExecutionMessage.h"
+#include "bcos-framework/interfaces/ledger/LedgerInterface.h"
+#include "bcos-framework/interfaces/protocol/BlockHeaderFactory.h"
+#include "bcos-framework/interfaces/protocol/TransactionReceiptFactory.h"
+#include "bcos-framework/interfaces/storage/StorageInterface.h"
 #include "mock/MockExecutor.h"
 #include "mock/MockExecutor3.h"
 #include "mock/MockLedger2.h"
