+++ libinitializer/SchedulerInitializer.h
@@ -19,16 +19,16 @@
  * @date 2021-06-21
  */
 #pragma once
-#include "interfaces/protocol/BlockFactory.h"
-#include "libinitializer/ProtocolInitializer.h"
-#include "libprotocol/TransactionSubmitResultFactoryImpl.h"
+#include "bcos-framework/interfaces/protocol/BlockFactory.h"
+#include "ProtocolInitializer.h"
+#include "bcos-framework/libprotocol/TransactionSubmitResultFactoryImpl.h"
 #include <bcos-framework/interfaces/dispatcher/SchedulerInterface.h>
 #include <bcos-framework/interfaces/executor/ExecutionMessage.h>
 #include <bcos-framework/interfaces/ledger/LedgerInterface.h>
 #include <bcos-framework/interfaces/storage/StorageInterface.h>
 #include <bcos-framework/libexecutor/NativeExecutionMessage.h>
 #include <bcos-framework/libtool/NodeConfig.h>
-#include <bcos-scheduler/SchedulerImpl.h>
+#include <bcos-scheduler/src/SchedulerImpl.h>
 
 namespace bcos::initializer
 {
